Code Refactoring, Error Handling, DB Models, Authentication, and Logic Routes
